hamamatsu: add support for NDPI images > 4 GB
Addresses issue #174 (and possibly others), regarding the reading of
large NDPI files.  This issue stems from the fact that classical TIFF
(and by extension, NDPI) images only support up to 32 bit values for
tagged metadata.  However, due to the nature of whole slide image data,
it isn't uncommon for the size of an NDPI to exceed the 32 bit range.
This means that key metadata, such as the byte positions of image layers
or their size in bytes, may be too large to store in a traditional TIFF
IFD entry.

Currently OpenSlide relies on heuristics to determine the high bits of 64
bit addresses, which fail in some cases.  However, this is unnecessary,
as NDPI actually stores the high bits of the offset/value of each tag in
4 byte blocks immediately after the end of the IFD.

This fix modifies openslide-decode-tifflike.c to append these extra 4
bytes to each IFD entry's value/offset and, if necessary, modifies its
type to LONG8.

This fix also modifies openslide-vendor-hamamatsu.c to construct correct
restart marker addresses.  Currently only the values in TIFF tag 65426
are used, which are only the lower 32 bits of each address.  High bits
are stored in TIFF tag 65432, so these are now appended before mcu_starts
are calculated.

Signed-off-by: Nick Trahearn <n.a.trahearn@gmail.com>
Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>